Move Svelte to peer dependency and update versions
Removed Sveltekit entirely as this is just a
component library which (hopefully) can be used
with both Svelte 3 and 4. Have Sveltekit as a
dependency raised peer dependency conflicts as
Sveltekit 2 is not compatible with Svelte 3 and
Sveltekit 1 is not compatible with Svelte 4.

Also updated a couple of other package versions to
resolve other peer dependency conflicts that arose
from updating the main package versions.